SimpleSolvers

SimpleSolvers.ArmijoQuadratic

SimpleSolvers.Bisection

SimpleSolvers.Options

SimpleSolvers.derivative

SimpleSolvers.derivative

SimpleSolvers.derivative!

SimpleSolvers.derivative!!

SimpleSolvers.gradient

SimpleSolvers.gradient

SimpleSolvers.gradient!

SimpleSolvers.gradient!!

SimpleSolvers.value

SimpleSolvers.value

SimpleSolvers.value

SimpleSolvers.value

SimpleSolvers.value!

SimpleSolvers.value!

SimpleSolvers.value!!

SimpleSolvers.value!!

SimpleSolvers.ArmijoQuadratic — Type

Quadratic Armijo line search

SimpleSolvers.Bisection — Type

simple bisection line search

SimpleSolvers.Options — Type

Configurable options with defaults (values 0 and NaN indicate unlimited):

x_abstol::Real = 1e-50,
x_reltol::Real = 2eps(),
f_abstol::Real = 1e-50,
f_reltol::Real = 2eps(),
g_abstol::Real = sqrt(eps()),
g_reltol::Real = sqrt(eps()),
x_abstol_break::Real = Inf,
x_reltol_break::Real = Inf,
f_abstol_break::Real = Inf,
f_reltol_break::Real = Inf,
g_abstol_break::Real = Inf,
g_reltol_break::Real = Inf,
f_calls_limit::Int = 0,
g_calls_limit::Int = 0,
h_calls_limit::Int = 0,
allow_f_increases::Bool = true,
min_iterations::Int = 0,
max_iterations::Int = 1_000,
warn_iterations::Int = max_iterations,
show_trace::Bool = false,
store_trace::Bool = false,
extended_trace::Bool = false,
show_every::Int = 1,
verbosity::Int = 1

SimpleSolvers.derivative!! — Method

Force (re-)evaluation of the derivative of the objective at x. Returns f'(x) and stores the derivative in obj.D

SimpleSolvers.derivative! — Method

Evaluates the derivative of the objective at x. Returns f'(x) and stores the derivative in obj.D

SimpleSolvers.derivative — Method

Evaluates the derivative of the objective at x. Returns f'(x), but does not store the derivative in obj.D

SimpleSolvers.derivative — Method

Get the most recently evaluated derivative of the objective of obj.

SimpleSolvers.gradient!! — Method

Force (re-)evaluation of the gradient at x. Returns ∇f(x) and stores the value in obj.g.

SimpleSolvers.gradient! — Method

Evaluates the gradient at x. Returns ∇f(x) and stores the value in obj.g.

SimpleSolvers.gradient — Method

Evaluates the gradient at x. This does not update obj.g or obj.x_g.

SimpleSolvers.gradient — Method

Get the most recently evaluated gradient of obj.

SimpleSolvers.value!! — Method

Force (re-)evaluation of the objective at x. Returns f(x) and stores the value in obj.F

SimpleSolvers.value!! — Method

Force (re-)evaluation of the objective value at x. Returns f(x) and stores the value in obj.F

SimpleSolvers.value! — Method

Evaluates the objective at x. Returns f(x) and stores the value in obj.f

SimpleSolvers.value! — Method

Evaluates the objective value at x. Returns f(x) and stores the value in obj.F

SimpleSolvers.value — Method

Evaluates the objective value at x. Returns f(x), but does not store the value in obj.f

SimpleSolvers.value — Method

Get the most recently evaluated objective value of obj.

SimpleSolvers.value — Method

Evaluates the objective value at x. Returns f(x), but does not store the value in obj.F

SimpleSolvers.value — Method

Get the most recently evaluated objective value of obj.